(wip) fixes to structure
