+++ src/main/java/frc/robot/constants/SimConstants.java
@@ -11,4 +11,7 @@ public class SimConstants {
 	public static final double STEER_INERTIA_KGMS2 = 0.05;
 
 	public static final double SIM_LOOP_PERIOD = 0.05; // 5 ms
+
+	public static final double STARTING_POS_X_FT = 2.3;
+	public static final double STARTING_POS_Y_FT = 3.2;
 }
